riskyurl-Github/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "cors>=1.0.1",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "pyopenssl>=25.0.0",
    "requests>=2.32.3",
    "reportlab>=4.2.5",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["script"]

